requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.22
pypdf==3.17.4
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers